feat: extract shared logic from NewLogicBlock to EditLogicBlock
to be shared between upcoming ActiveLogicBlock and NewLogicBlock component

also also move EditCondition components into EditLogicBlock
