+++ frontend/src/features/admin-form/create/logic/components/LogicContent/EditLogicBlock/index.ts
@@ -1,0 +1,1 @@
+export * from './EditLogicBlock'
